common: move WeightedRandomSelection one level up in package hierarchy
